extra/py/scripts/venv_requirements.txt
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
firebase-admin==4.3.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
gsheets==0.4.1
html2markdown==0.1.7
html2text==2020.1.16
httplib2==0.18.1
idna==2.10
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.10
pipreqsnb==0.2.3
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==20.0.0
qtconsole==5.0.0
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.2.1
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
zipp==3.4.0
